Add Axon framework
After discussion with @dualspiral, it's been decided that the architecture
of the application will be an Event Sourcing + Command Query Responsibility Segregation
(esCQRS for short). In the Java world, there are a few frameworks that allow
for setting up such an application, but finding one that will work in a gradle
structure is narrowing our choices. Axon provides enough of the framework
requirements to set up an esCQRS app, along with extensibility to make the
app have some more customizations than "just another spring app".

For now, we'll be starting with Spring for getting the application off the ground,
while later we migrate off from Spring, retaining the power of esCQRS.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>