Re-implement Synchronizer as a true Akka Streams actor
This is a pretty slick piece of work that takes advantage of Akka's actor model and Akka's
Streams+Clustering to do the various pieces of work together. The difference prior may
have been from the lack of control on parallelization to responding to Kafka messages,
but now, the actors perform various bits of work as expected.