Migrate entirely to Postgres
Started refactoring all of Version querying to a separated service.

There's some troubles here basically where the views are not quite yet
usable for querying, and the actual queries are a little finicky. It is
clear enough that JPA is useful for just dumping plain old data, but
using it to query data is a royal pain (hence the reason why we should
be splitting the read from writes).

I've been experimenting with the actual view models in IDEA and the
JPA console, so what may end up happening is just a devolved form of
using JPA to simply perform queries as needed, but not using it's
relationships entirely as currently written.