Make use of liquibase to set up the database
Locally, one can set up liquibase with the postgres server by
running the goal liquibase:update. I've swapped out the read
from the write models and forced JPA to only care about reading
from a materialized view that is refreshed as a trigger.

There are some advantages to this: we no longer have to write
the update logic, the database will do this a lot faster than we
can with transactions and such. There may be the caveat that
during the initial startup and "ingesting the world", the
materialized view may end up needing to be converted to a normal
table that gets a function that upserts the replacement rows.