Reimplement regex-based version recommendation
At this point, most all the functionality previously provided by the
old 'just query the current persisted entity' for state instead of
querying the database for the most-up-to-date-data is migrated. The last
bit of functionality required to migrate is manual version
recommendation.

After this, the fun of gathering artifact components from maven begins.
Fortunately, a lot about using Akka and workers has been learned up to
this point that enables writing the remaining bits of the system to be
better handled from scratch and less fumbling around.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>