Restore recommended version functionality
Lots of small/large changes to re-calculate recommended versions
based on the registered regex. Notably, the worker now fully
refreshes the particular artifacts and/or the materialized view on
the desired occasion. This does have the "eventual-consistency"
feel when querying the VersionQueryService while the synchronizer
does its initial work (some responses may have partial tags populated).

Now, the nice thing is that recommended versions are showing up as
expected in all responses, regardless whether a tag is filtered or
not.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>