Replace SonatypeClient with actor behaviors
It never felt right that the actor was just blindly calling some
blocking work to do, and potentially exceptionally not even handling the
failure cases. Now the behavior is set up to be handled correctly where
mapped results (or exceptioned results) are logged and after completion,
the ActorRef is told about the results. This does pose some interesting
problems with regard to "what do we do with failing artifacts",
but ultimately, this will suit us fine.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>