Enable configuration of the Synchronizer
Basically almost all the pools are configurable for timeouts, pool size,
parallelism, etc. to further enable tuning of the system however need be
for production/staging/deployed environments that aren't running in dev.
For the mean time, these settings are suitable for in-development, but I
can imagine that they may be tweaked for higher throughput/handling when
in containers.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>