Terraform more things. Monitoring sorta kinda works.
Doing more reading to understand how to set up lagom services with Akka
Discovery between pods to form a cluster. Ideally the pods themselves
will be small in memory footprint, and hold no persistent volumes. There
might be more to actually do with the docker image that's built for the
services themselves, and an addition of a maven dependency for
Akka to interact with the Kubernetes API.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>