Refactor commit resolution workflow
This is much more reactive than previously designed,
but the overall result leads for a cleaner separation of
concerns regarding what versions have a commit registered, what
repositories are available for an artifact, and whether a
version needs work to be done.

After all of this, we should have some better commit messages available.